make simplify module
